+++ projects/vdk-plugins/airflow-provider-vdk/.airflow-ci.yml
@@ -39,6 +39,7 @@ release-airflow-provider-vdk:
     - cd projects/vdk-plugins/
     - echo "Releasing airflow-provider-vdk..."
     - cd airflow-provider-vdk/ || exit 1
+    - sed -ri "s/__version__ = \"([[:digit:]]\.[[:digit:]]\.).*\"$/__version__ = \"\1${VDK_PATCH_VERSION}\"/g" setup.py
     - pip install -U pip setuptools wheel twine
     - python setup.py sdist --formats=gztar
     # provide the credentials as Gitlab variables
